requirements.txt
streamlit==1.40.2
boto3==1.35.78
python-dotenv==1.0.1
tavily-python==0.5.0
beautifulsoup4==4.12.3
requests==2.32.3